added small feature to the (default) column view
Added (hardcoded because quickhack) config parameter to tell the plugin to always list all pages unter their beginning letter before breaking into a new column. This feature may come in handy in some cases and could perhaps be extended with somekind of threshold for a hard cap of entries before breaking. If you merge this feature you should embed it with a parameter to enable/disable it within <nspage> tag