Update selenium
Beside updating the version in the pom we nowneed to ensure the geckodriver is
available. Hence an update of the setup script

Also, almost all tests failed after this update. It seems to be because the
execution of the test goes on even before the page completely loads. Hence the
ugly decorator, which adds sleeps and retries, to ensure tests pass.