Make the previous commit compatible with the option -r
That's because the pattern on which we match is `r *=? *\"?([[:digit:]]*)\"?`,
because we want to match for instance against all of this: `-r`, `-r=1`, `-r   =   1`.
Consequently `$matched` is `-r `, with already a trailing whitespace. So with the
added white space from your patch, it does not match anymore (and the page ends up
displaying `this namespace doesn't exist: -r` because at the end of the parsing the
`-r` is not removed from the parsed string).