Add $item parameter to nspages_printer::_printElementOpen
I believe this is how this function was intended to work, but `$item['type'] !== 'd'` would always evaluate to true, since there is no `$item` parameter. I added it at the end as an optional parameter to not break the signature. (However it is only called in `printerTree.php` so it could easily be changed.)